extract autocompletion in separate component
