+++ components/AddressSearch.tsx
@@ -22,12 +22,15 @@ import {
     closePanel
 } from '@/stores/map/slice';
 
+import AddressAutocomplete from '@/components/AddressAutocomplete'
+
 export default function AddressSearch() {
     const unknown_rnb_id = useSelector((state) => state.addressSearch.unknown_rnb_id)
 
     // URL params
     const params = useSearchParams()
     const [query, setQuery] = useState(params.get('q') || '')
+    const [keyDown, setKeyDown] = useState(null)
 
     // contains the address suggestions given by the BAN API
     const [addressSuggestions, setAddressSuggestions] = useState([])
@@ -43,31 +46,14 @@ export default function AddressSearch() {
     const apiUrl = 'https://api-adresse.data.gouv.fr/search/'
     const addressInput = useRef(null)
 
-    const handleKeyDown = async (e) => {
-        // called every time a key is pressed with search input focused
+    const handleKeyDown = (e) => {
         dispatch(setAddressSearchUnknownRNBId(false))
-        // pick the suggestion with the arrow keys
-        if (e.key === 'ArrowDown') {
-            e.preventDefault()
-            if (selectedSuggestion < addressSuggestions.length - 1) {
-                setSelectedSuggestion(selectedSuggestion + 1)
-            }
-        } else if (e.key === 'ArrowUp') {
-            e.preventDefault()
-            if (selectedSuggestion > 0) {
-                setSelectedSuggestion(selectedSuggestion - 1)
-            }
-            // select the suggestion with the enter key
-        } else if (e.key === 'Enter') {
-            if (queryIsRnbId()) {
-                dispatch(closePanel())
-                handleBdgQuery()
-            } else if (addressSuggestions.length > 1 && selectedSuggestion >= 0) {
-                select_suggestion(addressSuggestions[selectedSuggestion])
-                // you don't need to select a suggestion if there is only one
-            } else if (addressSuggestions.length == 1) {
-                select_suggestion(addressSuggestions[0])
-            }
+        if (e.key === 'Enter' && queryIsRnbId()) {
+            // special case, if the query is a RNB ID we bypass the address search
+            dispatch(closePanel())
+            handleBdgQuery()
+        } else {
+            setKeyDown(e)
         }
     }
 
@@ -80,13 +66,14 @@ export default function AddressSearch() {
         if (queryIsRnbId()) {
             dispatch(closePanel())
             handleBdgQuery()
-        } else {
-            handleAddressQuery().then((results) => {
-                if (results.length > 0) {
-                    select_suggestion(results[0])
-                }
-            })
-        }
+        } 
+        // else {
+        //     handleAddressQuery().then((results) => {
+        //         if (results.length > 0) {
+        //             select_suggestion(results[0])
+        //         }
+        //     })
+        // }
     }
     
     const handleBdgQuery = async () => {
@@ -125,49 +112,39 @@ export default function AddressSearch() {
         return mapPosition
     }
 
-    const handleAddressQuery = async () => {
-        // Add the query to the store
-        const geocode_result = await fetchBanAPI(addressInput.current.value);
-        dispatch(setAddressSearchQuery(addressInput.current.value))
-        dispatch(setAddressSearchResults(geocode_result.features))
-        if (geocode_result.features.length > 0) {
-            setAddressSuggestions(geocode_result.features)
-            setSelectedSuggestion(-1)
-        }
-        return geocode_result.features
+    // const handleAddressQuery = async () => {
+    //     // Add the query to the store
+    //     const geocode_result = await fetchBanAPI(addressInput.current.value);
+    //     dispatch(setAddressSearchQuery(addressInput.current.value))
+    //     dispatch(setAddressSearchResults(geocode_result.features))
+    //     if (geocode_result.features.length > 0) {
+    //         setAddressSuggestions(geocode_result.features)
+    //         setSelectedSuggestion(-1)
+    //     }
+    //     return geocode_result.features
 
-    }
-    const fetchBanAPI = async (query) => {
-        let query_url = new URL(apiUrl);
-        query_url.searchParams.set('q', query);
-        query_url.searchParams.set('autocomplete', 1);
-        return new Promise((resolve, reject) => {
-            fetch(query_url)
-                .then(response => response.json())
-                .then(data => {
-                    resolve(data)
-                })
-                .catch(error => {
-                    reject(error)
-                })
-        })
-    }
+    // }
+    // const fetchBanAPI = async (query) => {
+    //     let query_url = new URL(apiUrl);
+    //     query_url.searchParams.set('q', query);
+    //     query_url.searchParams.set('autocomplete', 1);
+    //     return new Promise((resolve, reject) => {
+    //         fetch(query_url)
+    //             .then(response => response.json())
+    //             .then(data => {
+    //                 resolve(data)
+    //             })
+    //             .catch(error => {
+    //                 reject(error)
+    //             })
+    //     })
+    // }
 
     useEffect(() => {
         if (params.get('q') !== null) {
             search()
         }
     }, [])
-
-    useEffect(() => {
-        setSelectedSuggestion(-1)
-        if (query.length < 3) {
-            setAddressSuggestions([])
-        } else if (!suggestionChosen) {
-            handleAddressQuery()
-        }
-        setSuggestionChosen(false)
-    }, [query])
 
     const select_suggestion = (suggestion) => {
         setAddressSuggestions([])
@@ -186,11 +163,11 @@ export default function AddressSearch() {
         })
     }
 
-    const suggestions = addressSuggestions.map((s, i) =>
-            <div onMouseEnter={() => setSelectedSuggestion(i)} onClick={() => select_suggestion(s)} className={styles.suggestion + ' ' + (selectedSuggestion == i ? styles.selected : '')} key={s.properties.id} >
-                {s.properties.label}
-            </div >
-    );
+    const handleSuggestionSelected = ({ suggestion }) => {
+        if (suggestion !== null) {
+            select_suggestion(suggestion)
+        }
+    }
 
     return (
         <>
@@ -207,9 +184,9 @@ export default function AddressSearch() {
                     ref={addressInput}
                     onKeyDown={handleKeyDown}
                 />
-                {addressSuggestions.length > 0 && <div className='fr-pt-1v'>{suggestions}</div>}
-            </div>
 
+                <AddressAutocomplete query={query} keyDown={keyDown} onSuggestionSelected={handleSuggestionSelected} ></AddressAutocomplete>
+            </div>
         </>
     )
